Catalog: Ensure epoch does not decrease
Previously migrating from one catalog implementation to another
allowed the catalog epoch to decrease. The problem was that we were
treating the epoch as an internal implementation detail. During the
migration from the stash to persist, we would NOT copy the epoch from
the stash to persist. Instead, persist would start with a brand-new
epoch at 2 (the starting value is 2 for historical reasons). The
problem is that the epoch is exposed externally, and it's used by the
storage controller. The storage controller uses the epoch for its own
separate fencing purposes. If the epoch decreases after a migration,
then the storage controller would mistakenly think it was being fenced
out by a higher epoch and fail to boot.

This commit adds an epoch lower bound to catalog opening. When the
catalog opens, it ensures that it's epoch is at least as big as the
epoch lower bound. During a migration, we pass the old epoch plus one
as the epoch lower bound, which ensure that the epoch never decreases.

Works towards resolving #24218