adapter: base read policies on implied_capability
This commit adjusts adapters read policy code to always use collections'
`implied_capability`s, rather than `read_capabilities.frontier()`.

This fixes a potential bug where adapter might try to install a read
hold on a collection based on `read_capabilities.frontier()` through a
`ValidFrom` read policy but fail to do so because the read hold time is
before the collection's `implied_capability`.

Note that it is not clear whether the described potential bug can ever
occur in Materialize. Failures to regress the `implied_capability` when
installing a read policy are silently ignored, and it's not easy to
change that because some parts of the code assume that it is valid to
set a `ReadPolicy::ValidFrom([0])` to disable compaction of a
collection regardless of its current `implied_capability`.