catalog: Read LD before catalog open
This commit moves the sync with LaunchDarkly on startup from after
opening the durable catalog to before opening the durable catalog. This
allows us to use LaunchDarkly flags to determine how to open the
durable catalog. Specifically this allows brand-new environments to use
the catalog implementation specified in LaunchDarkly, without needing
to reboot.

Works towards resolving #24218