compute: add `Context::region_named`
This commit extens the `Context` type with an API to create dataflow
regions that get their own `Context`. The idea is that every rendering
function will execute in a context that has been moved to the
appropriate scope, so type annotations become simpler and we can
consolidate `enter_region`/`leave_region` calls.