compute: clean up join dispatch calls
This commit cleans up the dispatch calls in join rendering code by
simplifying the turbofish type annotations. Those are necessary because
of compiler limitations, but they don't have to be as extensive as they
previously were. Replacing most of them with `_` wildcards makes the
code less boilerplate-y and also unbreaks rustfmt in one instance.