adapter: fix tracing span plumbing
Looks like this was accidentally severed in #23577.

Turns out we have a test, there was just a "bug" in the test where it
was asserting things worked with tracing turned on for "debug", which
they did, but that's somewhat uninteresting because we can only turn
tracing to always-on at "info". This also updates the test, which fails
without this fix.