sql: error when referencing ungrouped column that also exists in outer scope
Our existing code for detecting references to columns that did not
appear in the GROUP BY clause or an aggregate function did not consider
the case that a column of the same name existed in the outer scope. Per
PostgreSQL, the correct behavior is to bind to the column reference in
the inner scope and error due to the lack of grouping, rather than to
successfully bind to the column reference in the outer scope.

Fix #23748.
Fix #24328.

Co-authored-by: Rainer Niedermayr <rainer@materialize.com>