persist-txn: fix mem leak in TxnsRead
`TnxsRead` was not previously compacting its internal representation as
time advanced, which was a memory leak. It only needed the historical
distinctions to serve `data_listen_next` queries, but that method was
as-yet unused. We might want to add it back, but for now, fix the leak
by removing it and then compacting whenever the progress advances.