Consolidate collections without building batches
This changes consolidation to stream output without building intermediate
batches. Ideally, it avoids a memory spike that's caused by keeping around
two representations of the same data.

The implementation buffers the output in appropriately-sized batches as
expected by Timely, but it only surfaces them after the whole batch's data
is ready. An alternative implementation could directly surface the data,
which could allow some downstream operators to see the data earlier. This
is currently not possible due to limitations in the types used by the
builder.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>