catalog: Fix tombstone value for emergency stash (#24376)
The emergency stash implementation is an escape hatch which allows us to
use the stash catalog in a customer's environment without performing a
rollback. It's meant for emergency situations when a rollback is
impossible. Previously, when using the emergency-stash Materialize would
not update the catalog tombstone value in the stash. The tombstone value
is used to determine which catalog implementation to use during start
up. As a consequence Materialize could end up in a situation where
Materialize looks at the wrong catalog during start up. For example:

  1. The tombstone value is true, i.e. use the persist catalog.
  2. Start the environment with the emergency stash catalog.
3. Make changes to the catalog in the stash that are expected to be
visible on restart.
  4. Restart.
5. The tombstone value is still true, so perform all reads against
persist during startup.

This commit updates the emergency stash implementation such that it
ensures the tombstone value is false, which causes materialize to read
from the stash during start up.

Works towards resolving #24218