persist-txn: fix optimizer cost statistics for lazy
This is yet another example of a read path needing to learn how to go
through persist-txn for shards that it manages because the physical
upper may be lagging behind the logical upper. The other ones all use
the `unblock_read` trick but we don't want to do this here because it's
too slow (this happens as part of planning). Instead, use the "translate
as_of" trick that we invented previously but didn't end up using.