test: use explicit clusters in source and sink testdrive files
we recently deprecated linked clusters and began using the session's
default cluster when no cluster's specified for sources and sinks.
this has the annoying side effect that if the default cluster uses
more than one replica, the source or sink fails--failing the test.

rather than hunt down which test expects which behavior, just use
clusters in sources and sinks in testdrive files.