adapter: Fix timeline determination for temporals
Previously we would determine the timeline context of a query by
performing BFS over all the objects in the query and their
dependencies. At each dependency we would add a timeline using the
following logic:

  - If the current object is a leaf view, and we HAVE previously seen a
  view with a temporal function then add the `TimestampDependent`
  timeline context.
  - If the current object is a leaf view, and we HAVE NOT previously
  seen a view with a temporal function then add the
  `TimestampIndependent` timeline context.
  - All other objects directly add their timeline.

The final timeline contexts from this process is dependent on the
traversal order of each node. Specifically if a view with a temporal
function is visited after all leaf views have been visited, then we
will incorrectly not add any `TimestampDependent` contexts.

This commit fixes this issue by changing the logic to the following:

  - If the current view contains a temporal function then add the
  `TimestampDependent` timeline context.
  - If the current view does not contain a temporal function then add
  the `TimestampIndependent` timeline context.
  - All other objects directly add their timeline.