persist: replace PersistFeatureFlag with new dyncfg system
This migrates the three PersistFeatureFlags with the equivalent Configs
in the new dyncfg system and then does the necessary hookup to make this
work end-to-end. The definitions of the flags/configs are moved out of
cfg.rs to be next to the usages to establish the new convention.

We'd also like to replace the rest of DynamicConfig with this (and
potentially start swapping over non-persist parts of the codebase) but
this at least establishes the plumbing.