compute: make `Context` a trait
This commit replaced the `Context` struct with a trait, in preparation
of having different contexts corresponding to different dataflow scopes.
This requires making the previous `Context` methods free-standing
functions because trait method definitions cannot be split across
modules like regular method definitions can.

Here we only add a `RootContext` and the all the fixing up necessary for
converting methods to functions. Actual nested contexts are introduced
in a subsequent commit.