sql: add maybe_grow to NameResolver::fold_function, fold_expr
The implementation of our reg* types relies on heavily nested
queries--when resolving the names for those functions on the
smaller threads available outside of the coordinator's main loop
(e.g. when opening the catalog), we were overflowing our stack.

Because `impl Fold` does not allow early returns, our easiest answer
is to just grow the stack size.