persist-txn: use real timestamp oracle in maelstrom tests
This hooks up the production impl of ShareableTimestampOracle to the
persist-txn maelstrom tests. It also adds an in-mem impl of that trait
for convenience when running a single-node test.

It should be possible to also make the MaelstromOracle implement this
trait (I think I have it working locally) but it's just tricky enough
that I didn't feel great about it and seems like that doesn't need to
hold up this PR (we just lose the fancy graphs).