compute: introduce child contexts
This commit introduces a `ChildContext` type to represent rendering
contexts inside dataflow regions. It then adjusts all rendering methods
that take both a (global) context and a scope to instead only take a
context of the appropriate scope.

The result is that we can get rid of a bunch of scope type variables,
simplifying function prototypes. Plus, this pattern forces us to pull
`enter_region`/`leave_region` calls into the function that creates the
region rather than the one that renders its contents. This improves
readability by separating concerns better.