rqg: Pass --duration=0 to the RQG as well
Previously, the RQG was not even run if the mzcompose was called
with --duration=0. This means that any initialization steps
that are in the grammar will not get run.

Remove the special-case for --duration=0, so that the RQG is called,
It will have the opportunity to run the grammar initialization and
will then terminate immediately.