test: allow waiting longer when verifying slot
when verifying the state of a slot, we could hit some unfortunate
timing when dropping the slot on a persistent cluster (i.e. one
that will continue running after the source is dropped). namely,
when dropping a source, we very quickly spawn a task to drop
the replication slot, while communicating with the cluster that
the dataflow needs to be dropped takes slightly longer. if the task
dropping the replication slot is unfortunate, it's possible that its
retry loop backs off too quickly to effectively be retried after we
know that the dataflow has stopped being rendered.

these kinds of race conditions between the adapter and storaged are
well known and are actively being worked on. However, until we make
the ordering of commands deterministic, the easiest thing to do is
just wait longer for the dataflow to stop being rendered and
let us drop the replication slot.