storage/pg: set PG source to starting on render
many users are confused that PG sources' primary sources state
that they're starting instead of running, even while the ingestion
is running. we can patch over this by just seeding the state of the
primary source with a Running update.

Note that it is expected behavior that restarting the source
will lose definite errors in the status collection. #23950