Improve spines in top-k arrangement stacks
This commit improves further the spine types employed in the
arrangement stacks built for top-k computation. Previously, only the
key type could benefit from DatumContainer representation. With this
change, now value types can also employ this more advantageous
container type in all levels of the stack but the first. This is
because the first level of the stack needs to represent Result
instances to track invalid accumulation errors.