persist: Allow read-only users to connect to persist (#24474)
Previously, connecting to persist required a CRDB user with enough
privileges to alter the `consensus` table in CRDB. The ALTER statement
is meant as a performance improvement and is a no-op for every user
except the first user that connects to persist. This requirement
prevented users who only wanted to read from persist and only had CRDB
users with SELECT privileges from connecting to persist. Connecting as a
read-only user would be especially useful for the catalog-debug and
persist-cli tools.

This commit updates the persist connection process, so that CRDB users
without the privileges to run ALTER can still connect and read from
persist.