compute-client: add an explicit drop flag for collections
So far the compute controller considered a collection dropped when its
read frontier has advanced to the empty frontier. This had the defect
that we might remove a collection's state from the controller in
response to a client setting a `ReadPolicy::ValidFrom([])`. A client
might still expect to be able to query information about a collection
even when it is not interested in reading data from it anymore.

To remove this defect, a `CollectionState::dropped` flag is introduced
and used to explicitly mark collections as dropped when
`drop_collections` is called. Only collections that have the `dropped`
flag set are then considered for having their controller state dropped.

Figuring out the right places in which to check the collection state
removal conditions becomes a bit of a hassle with this change, so we
also rearrange things so that the collection state cleanup is performed
asynchronously in a period task. This makes us keep unneeded collections
around a bit longer than necessary but also greatly simplifies the rest
of the code.