[200~adapter: Always attempt to sync LD on boot
Previously, starting up an environment would only attempt to sync with
LaunchDarkly if the environment had never before synced with
LaunchDarkly. Otherwise, the environment would wait until startup has
completed before attempting to sync with LaunchDarkly. If an
environment was stuck in a crash loop unable to fully startup, and the
underlying reason was the value of some system variable, then we would
be unable to resolve the issue without manually editing the catalog.

This commit updates the startup process, so that if an environment has
already synced with LaunchDarkly, then it still attempts to sync during
startup, but will time out after 30 seconds. This change will allow us
to update system variables of environments that are stuck in a crash
loop.

Works towards resolving #24316