catalog: Fix upgrade check during migrations (#24472)
Previously, if catalog implementation A was initialized, implementation
B was uninitialized, we tried to open a savepoint catalog using the
`CatalogMigrator` implementation, and our target implementation was B,
then the savepoint would fail to open. The savepoint catalog is unable
to work with uninitialized catalogs and is therefore unable to perform
an upgrade check when a migration/rollback is needed. Instead of failing
startup, this commit will not attempt to perform an upgrade check in
this scenario.

Works towards resolving #24218