carry `plan: SelectPlan` in all `PeekStage`s
This commit changes all structs that represent `PeekStage` variants to
carry `plan: SelectPlan` field instead of its deconstructed constituent
parts.

The rationale is that (a) the `EXPLAIN` path finalization stage (to be
added shortly) might need fields that are otherwise pruned out by
earlier stages, and (b) the developer burden of threading through new
fields that are needed in the final stage through all intermediate
stages will be reduced as long as the field is part of the `SelectPlan`.