catalog: Use cache when fetching epoch in startup
Previously, when the persist catalog would fetch the catalog epoch
before opening the catalog, it would completely ignore the snapshot
cache and fetch an entire snapshot directly from persist every time.
This is wasteful because we have a snapshot cache available in memory
and don't need to go to persist every time.

This commit updates the persist catalog to look at the snapshot cache
when fetching the epoch.

Works towards resolving #24218