adapter: panic on more controller errors
This commit adjusts the `ShouldHalt` implementations of controller
errors to be less lenient. All of the errors that panic now are ones
that we don't expect to see in the absence of bugs, so if they are
returned we want to process to panic so we get notified in Sentry.