adapter: allow forcibly dropping replication slot
Previously, we could leak replication slots in upstream PG
databases if the storage cluster kept the replication connection
alive longer than expected.

To mitigate this, introduce a method of forcibly stopping the
replication slot's backend process. This is safe to do via
pg_terminate_backend according to PG's docs:

> Terminates the session whose backend process has the specified
> process ID. This is also allowed if the calling role is a member
> of the role whose backend is being terminated or the calling
> role has privileges of pg_signal_backend, however only superusers
> can terminate superuser backends.

https://www.postgresql.org/docs/current/functions-admin.html

We attempt to do this gracefully before resorting to terminating
the backend because we should be able to drop the replication slot,
barring network issues. If we succeed in forcefully dropping the
replication slot, log an error so that we are alerted of its
occurrence because we don't expect this code path to execute
successfully, but also don't want to leak resources due to bugs
in our platform.