catalog: Consolidate transactions in memory
This commit will consolidate durable catalog transactions in memory
before committing them durably. Consolidating in-memory is usually much
faster than consolidating on disk, so we optimistically do so. This is
especially true of the stash which can fall over if too many
retractions are written. This is even more true during catalog
rollbacks form persist to the stash which may write an extremely large
number of retractions that can all be consolidated in memory.