persist: switch the remaining Vars configs to the new dyn config system
Added two new `ConfigType`s: `u32` and `Duration`.

It turns out that a few DynamicConfigs were never actually hooked up to
vars for dynamic configuration. Leaving these for a followup, to try to
break this up a tiny bit. Anyway, we may want to make everything in
PersistConfig dynamically configurable now that it's easy, and these
would fit pretty naturally in that category.

Mostly mechanical modulo making sure that some of the configs are still
available externally to be overridden by tests outside of
persist-client.