compute-client: factor out a `ReplicaState`
This commit factors out a `ReplicaState` struct to contain all state
maintained by the `Instance` controller about a replica. This state was
previously mostly contained within the `Replica` struct that serves as a
client to a replica task and shouldn't be in the business of also
holding state for the use by the `Instance` controller. This commit also
renames the `Replica` type to `ReplicaClient`, to make the distinction
clearer.

The other piece of state that now resides inside `ReplicaState` is the
flag indicating whether a replica has failed. This was previously in a
`failed_replicas` map under `Instance`. Moving it into the per-replica
state frees us from having to remember to remove entries from
`failed_replicas` whenever we remove them from `replicas`.