compute-client: move hydration status tracking to `Instance`
This commit moves the hydration status tracking for compute collections
based on their frontiers from the `ReplicaTask` to the `Instance`
controller.

This prepares us for a world in which the compute controller allows
replicas to drop their compute resources for collections that are still
considered alive by the controller. In this world the controller might
send an `AllowCompaction([])` command for a collection even though that
collection has not yet been dropped, so we would like to keep around the
introspection data about its hydration status. However, the
`ReplicaTask` does not know about the drop status of collections since
it only observes compute commands, so it might drop the hydration status
data too soon. Moving the hydration status tracking to the `Instance`
solves this issue.