catalog: Fix shadow catalog reconciliation
The timestamp collection and the storage usage collection are updated
on a regular frequency. As a result, if the shadow catalog happens to
stop after updating one catalog but not the other, those two
collections can diverge. To fix this, the shadow catalog has a
reconciliation process on boot that fixes any discrepancies found in
these collections.

This commit fixes on off by one error in the storage usage collection
reconciliation and adds tests for both reconciliation processes.

Fixes #24523