storage/pg: add log correlating ID to replication slot
A few times when debugging PG source issues, I've had to add a log
correlating a replication slot to a GlobalId––there is nowhere
else that does this.

I'm adding the log here because the replication stream always
gets opened, whereas the snapshot stream only gets opened if there
are subsources to snapshot.