persist: migrate stats untrimmable columns to new dyn config system
The rest of them are also switched in the next commit, but pulling this
one out because there's not a great way to migrate it 1:1.

I've never loved that this is a json value that gets deserialized to a
struct with serde, so taking the opportunity to split it into three
configs, each a comma separated list. There no existing overrides in LD
to consider (in fact it hasn't even been created).

Another possibility for the dyn cfg migration would be to leave it as a
single string config containing json if someone feels strongly about
keeping this the way it worked before (though note that we'd lose the
adapter validation of the config value).