catalog: Restructure unopened persist catalog (#24519)
This commit restructures `UnopenedPersistCatalogState` to reduce the
amount of snapshots read from persist. Previously, all reads done by
`UnopenedPersistCatalogState` would require either getting an entire
snapshot from persist or looking at a saved cache if the timestamps
happen to match up. Now the `UnopenedPersistCatalogState` is able to
incrementally maintain a snapshot of the catalog by taking advantage
of the persist Listen API, which guarantees that we take a maximum of
one snapshot before opening the catalog.

Additionally, the `UnopenedPersistCatalogState` is able to pass along
its snapshot and listen handle to the `PersistCatalogState`, which
removes the need for an additional snapshot during open.

Works towards resolving #24218