persist-txn: fix optimizer cardinality estimates for "lazy"
When `latest_write` on `DataSnapshot` is `None`, it means persist-txn
knows that the latest write has been applied, but it has compacted away
information about specifically what timestamp that was. This makes it
hard to do the "as_of translation" for the `snapshot_stats` call. Work
around this by adding the ability to `snapshot_stats` to get whatever
the current stats are.