adapter: fix txn safety for INSERT RETURNING
If RETURNING is set in INSERT, some query execution occurs. INSERT can
usually only be used in single statement transactions for
serializability reasons. When we added RETURNING, we did not also
prevent those INSERTS from being executed in multi statement
transactions. They need to be prevented because we have already passed
the cluster auto routing and checks, which previously would cause a
panic because the RETURNING part was trying to execute on a different
cluster than the first query in the transaction.

Fixes #24612