adapter: use transaction cluster for restriction check
Previously we were using the `SET CLUSTER` cluster, even if auto routing
had previously occurred and already pinned the transaction to the
introspection cluster. This allowed user queries to be executed on the
introspection cluster is wrong.

Change the check method to use the same cluster as the transaction
instead of the one from the session vars.