compute: broadcast `UpdateConfiguration` commands
This commits makes the `PartitionedComputeState` broadcast
`UpdateConfiguration` commands to all shards, not just the first one.
The purpose is to ensure that all parts of the compute networking stack
are able to observe configuration updates.

Unfortunately, this makes the compute protocol more complex by
introducing another special case. We now have this situation:

 * `CreateTimely` is broadcast and accepted by all worker threads.
 * `UpdateConfiguration` is broadcast but only accepted by the first
   worker thread (and ignored by the others).
 * All other commands are only sent to the first worker thread.