catalog: Fix LargeCollectionStartupCache
Previously, the LargeCollectionStartupCache assumed that the cache was
append only and therefore when inserting new values ignored the diff.
For the most part this was true because audit events and storage usage
events are append only. However, during a catalog store migration or
rollback, we can actually retract audit events and storage usage
events. This commit teaches the LargeCollectionStartupCache how to deal
with retractions.

Works towards resolving #24218