deps: upgrade to rust-rdkafka with non-blocking drop
@guswynn discovered that dropping a Kafka consumer or producer blocks
until the consumer/producer is fully shut down (i.e., outstanding
messages flushed, offsets committed, rebalance assignment released).
This can cause deadlocks if the Kafka consumer or producer is dropped
from an async task that is scheduled on the same Tokio core thread as an
async task that is handling a librdkafka callback (e.g., establishing an
SSH tunnel).

This is particularly bad when dropping a Kafka consumer or producer from
`environmentd`, as we do during connection validation, as there are
50-50 odds that we deadlock on the thread handling the main coordinator
loop, thus stalling out the entire system.

The updated version of rust-rdkafka includes a patch to always drop
Kafka consumers/producers from a dedicated thread. Since this thread is
guaranteed not to be a Tokio thread, the drop will not deadlock.

This fixes the root cause of incident 92.