adapter: Fix timestamp selection of constant materialized views
Previously, a materialized view with a constant query without temporal
filters was considered timestamp independent, as in the adapter would
be free to choose any timestamp for the query. While it's true that the
timestamp selected would not affect the answer to the query, it would
affect our ability to query the materialized view. Materialized views
must durably materialize the results of their definition at a given
timestamp before we can query this timestamp. This commit updates the
timeline context logic to accurately reflect this.