adapter: acquire peek read holds before off-thread work
The recent optimize lir step was off-thread, and it acquired timestamps
and sometimes read holds before the spawn. During the spawn, the since
could advance past the determined timestamp and cause a panic. Read
holds were correctly acquired in the multi statement txn case, but
single statement txns were exempt from read holds because previously it
was guaranteed they were not needed since the timestamp was determined
in the same coord message that created the dataflows.

Now that there is off-thread work after determination, read holds must
always be acquired for all statements: single statements or `AS OF`
included.

See #24130

It's unclear if this fixes that issue completely since it was filed
before the optimize lir step. We should watch for more occurrences of
this bug in CI and close if they stop occurring.