adapter: refactor notice implmenetation to remove dep on CatalogState
CatalogState had an implementation for ExprHumanizer, which
largely duplicated the implementation on ConnCatalog. "Humanizing"
optimizer notices seemed to motivate the implementation; however,
the existing implementation on ConnCatalog serves the same purpose
more clearly.

For instance, when emitting notices to a session, using the
CatalogState impl of ExprHumanizer does not produce minimally
qualified names. Conversely, to guarantee consistent name resolution
for the system table, we want to use the system session ConnCatalog.