storage: remove kafka producer limits in sinks
The  kafka sink operator has nothing better to do with incoming data
other than to buffer them, which provides no additional value than to
send them to `librdkafka` and let that buffer them instead.

The operator is already set up to never buffer messages that are ready
to send but the current limits are too conservative. If a large snapshot
arrives fast enough then it is possible to reach the 10M message limit
and cause the sink to restart. I have observed this happening locally
during benchmarking.

For this reason this PR disables rdkafka message count and size limits.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>