adapter: remove now unnecessary append and group commit
As the comment describes, this append and group commit are intended to
make the rest of the tables immediately readable at the timestamp that
the new tables was added. With persist_txn now always on, this is
already guaranteed as part of the create_collections, so remove the
extra code and work.