compute-types: perform lowering within a `Context`
This commit introduces a `lowering::Context` type within which
MIR-to-LIR lowering is performed. The advantage of doing so is that
context information, such as the current bindings and feature flags,
don't need to be passed around as parameters through all recursive
method calls anymore. This will also make it easier to add additional
pieces of context in the future.