persist-txn: use TxnsRead worker task in operator
Previously, each time the persist-txn operator was rendered, we'd crate
a (filtered) subscribe to the txns shard. This replaces it with a
process-singleton (unfiltered) subscribe.

There is a notable tradeoff here between N filtered vs 1 unfiltered
subscriptions. However, in my staging env when there was a bug that held
back the txns since, it caused a _very_ :nervous-laughter: and linearly
increasing amount of cpu, memory, and network use. Because of this, I
believe it's worth selecting the option that bounds the worst case over
the option that optimizes the case of a few tables.

We could later get the "optimizes the case of a few tables" back, if
necessary, with a decent amount of elbow grease: making TxnsRead more
sophisticated about quiescing as well as dynamically changing the set of
data_ids that it watches.